+++ src/narrativelog/routers/add_message.py
@@ -28,19 +28,40 @@ async def add_message(
         default=[],
         description="Tags describing the message. " + TAG_DESCRIPTION,
     ),
-    systems: list[str] = fastapi.Body(
-        default=[],
+    systems: None
+    | list[str] = fastapi.Body(
+        default=None,
         description="Zero or more systems to which the message applies.",
     ),
-    subsystems: list[str] = fastapi.Body(
-        default=[],
+    subsystems: None
+    | list[str] = fastapi.Body(
+        default=None,
         description="Zero or more subsystems to which the message applies",
     ),
-    cscs: list[str] = fastapi.Body(
-        default=[],
+    cscs: None
+    | list[str] = fastapi.Body(
+        default=None,
         description="Zero or more CSCs to which the message applies. "
         "Each entry should be in the form 'name' or 'name:index', "
         "where 'name' is the SAL component name and 'index' is the SAL index.",
+    ),
+    components: None
+    | list[str] = fastapi.Body(
+        default=None,
+        description="Zero or more components to which the message applies. "
+        "Each entry should be a valid component name entry on the OBS jira project.",
+    ),
+    primary_software_components: None
+    | list[str] = fastapi.Body(
+        default=None,
+        description="Primary software components to which the message applies. "
+        "Each entry should be a valid component name entry on the OBS jira project.",
+    ),
+    primary_hardware_components: None
+    | list[str] = fastapi.Body(
+        default=None,
+        description="Primary hardware components to which the message applies. "
+        "Each entry should be a valid component name entry on the OBS jira project.",
     ),
     urls: list[str] = fastapi.Body(
         default=[],
@@ -98,10 +119,10 @@ async def add_message(
     tags = normalize_tags(tags)
 
     message_table = state.narrativelog_db.message_table
-
-    # Add the message.
+    jira_fields_table = state.narrativelog_db.jira_fields_table
     async with state.narrativelog_db.engine.begin() as connection:
-        result = await connection.execute(
+        # Add the message.
+        result_message = await connection.execute(
             message_table.insert()
             .values(
                 site_id=state.site_id,
@@ -122,6 +143,48 @@ async def add_message(
             )
             .returning(sa.literal_column("*"))
         )
-        result = result.fetchone()
+        row_message = result_message.fetchone()
+        if row_message is None:
+            raise fastapi.HTTPException(
+                status_code=http.HTTPStatus.INTERNAL_SERVER_ERROR,
+                detail="Couldn't create message entry",
+            )
 
-    return Message.from_orm(result)
+        # Add the jira fields
+        if any(
+            field is not None
+            for field in (
+                components,
+                primary_software_components,
+                primary_hardware_components,
+            )
+        ):
+            result_jira_fields = await connection.execute(
+                jira_fields_table.insert()
+                .values(
+                    components=components,
+                    primary_software_components=primary_software_components,
+                    primary_hardware_components=primary_hardware_components,
+                    message_id=row_message.id,
+                )
+                .returning(sa.literal_column("*"))
+            )
+            row_jira_fields = result_jira_fields.fetchone()
+
+            if row_jira_fields is None:
+                raise fastapi.HTTPException(
+                    status_code=http.HTTPStatus.INTERNAL_SERVER_ERROR,
+                    detail="Couldn't create jira_fields entry for message",
+                )
+
+        # Find the message and join with jira_fields
+        result_message_joined = await connection.execute(
+            message_table
+            # Join with jira_fields_table
+            .join(jira_fields_table, isouter=True)
+            .select()
+            .where(message_table.c.id == row_message.id)
+        )
+        row = result_message_joined.fetchone()
+
+        return Message.from_orm(row)